+++ packages/in_app_purchase/in_app_purchase_storekit/lib/src/in_app_purchase_storekit_platform.dart
@@ -153,11 +153,19 @@ class InAppPurchaseStoreKitPlatform extends InAppPurchasePlatform {
   }
 }
 
+enum _TransactionRestoreState {
+  notRunning,
+  waitingForTransactions,
+  receivedTransaction,
+}
+
 class _TransactionObserver implements SKTransactionObserverWrapper {
   final StreamController<List<PurchaseDetails>> purchaseUpdatedController;
 
   Completer? _restoreCompleter;
   late String _receiptData;
+  _TransactionRestoreState _transactionRestoreState =
+      _TransactionRestoreState.notRunning;
 
   _TransactionObserver(this.purchaseUpdatedController);
 
@@ -165,6 +173,7 @@ class _TransactionObserver implements SKTransactionObserverWrapper {
     required SKPaymentQueueWrapper queue,
     String? applicationUserName,
   }) {
+    _transactionRestoreState = _TransactionRestoreState.waitingForTransactions;
     _restoreCompleter = Completer();
     queue.restoreTransactions(applicationUserName: applicationUserName);
     return _restoreCompleter!.future;
@@ -176,6 +185,14 @@ class _TransactionObserver implements SKTransactionObserverWrapper {
 
   void updatedTransactions(
       {required List<SKPaymentTransactionWrapper> transactions}) async {
+    if (_transactionRestoreState ==
+            _TransactionRestoreState.waitingForTransactions &&
+        transactions.any((transaction) =>
+            transaction.transactionState ==
+            SKPaymentTransactionStateWrapper.restored)) {
+      _transactionRestoreState = _TransactionRestoreState.receivedTransaction;
+    }
+
     String receiptData = await getReceiptData();
     List<PurchaseDetails> purchases = transactions
         .map((SKPaymentTransactionWrapper transaction) =>
@@ -191,10 +208,21 @@ class _TransactionObserver implements SKTransactionObserverWrapper {
   /// Triggered when there is an error while restoring transactions.
   void restoreCompletedTransactionsFailed({required SKError error}) {
     _restoreCompleter!.completeError(error);
+    _transactionRestoreState = _TransactionRestoreState.notRunning;
   }
 
   void paymentQueueRestoreCompletedTransactionsFinished() {
     _restoreCompleter!.complete();
+
+    // If no restored transactions were received during the restore session
+    // emit an empty list of purchase details to inform listeners that the
+    // restore session finished without any results.
+    if (_transactionRestoreState ==
+        _TransactionRestoreState.waitingForTransactions) {
+      purchaseUpdatedController.add(<PurchaseDetails>[]);
+    }
+
+    _transactionRestoreState = _TransactionRestoreState.notRunning;
   }
 
   bool shouldAddStorePayment(