+++ script/tool/test/common/git_version_finder_test.dart
@@ -14,7 +14,7 @@ void main() {
   late List<List<String>?> gitDirCommands;
   late String gitDiffResponse;
   late MockGitDir gitDir;
-  String? mergeBaseResponse;
+  String mergeBaseResponse = '';
 
   setUp(() {
     gitDirCommands = <List<String>?>[];
@@ -74,7 +74,7 @@ file2/file2.cc
     final GitVersionFinder finder = GitVersionFinder(gitDir, null);
     await finder.getChangedFiles();
     verify(gitDir.runCommand(
-        <String>['diff', '--name-only', mergeBaseResponse!, 'HEAD']));
+        <String>['diff', '--name-only', mergeBaseResponse, 'HEAD']));
   });
 
   test('use correct base sha if specified', () async {