+++ packages/local_auth/CHANGELOG.md
@@ -1,3 +1,7 @@
+## NEXT
+
+* Updates example app Android compileSdkVersion to 31.
+
 ## 1.1.8
 
 * Update minimum Flutter SDK to 2.5 and iOS deployment target to 9.0.