+++ packages/local_auth/CHANGELOG.md
@@ -1,6 +1,6 @@
 ## NEXT
 
-* Updates example app Android compileSdkVersion to 31.
+* Updates Android compileSdkVersion to 31.
 
 ## 1.1.8